backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Cors
psycopg2-binary
google-generativeai
python-dotenv
pytest
